Pix.Env (pix v0.0.0)

Environment information.

Functions

arch()

@spec arch() :: Pix.Config.supported_arch() | (forced_arch :: atom())

Returns the current system architecture. Detects amd64/x86_64 or arm64/aarch64 architectures. Can be overridden by setting PIX_FORCE_PLATFORM_ARCH environment variable.

ci?()

@spec ci?() :: boolean()

Returns whether the code is running in a CI environment by checking the CI environment variable. Returns true if CI env var is set, false otherwise.

git_commit_sha(default \\ "")

@spec git_commit_sha(default :: String.t()) :: String.t()

Returns the current git commit SHA as a string. Executes git rev-parse HEAD command. Defaults to the given default value if git is not initialized or the command fails.

git_project_name(default \\ File.cwd!())

@spec git_project_name(default :: String.t()) :: String.t()

Returns the git project name (repository name) as a string. Gets the basename of the git root directory. Defaults to the given default value if git is not initialized or the command fails.

groupid()

@spec groupid() :: String.t()

Returns the current group ID by executing id -g.

os()

@spec os() :: String.t()

Returns the operating system name by executing uname -s.

pix_docker_build_opts()

@spec pix_docker_build_opts() :: OptionParser.argv()

Returns additional Docker build options from PIX_DOCKER_BUILD_OPTS environment variable. Parses the options string into command line arguments.

pix_docker_run_opts()

@spec pix_docker_run_opts() :: OptionParser.argv()

Returns additional Docker run options from PIX_DOCKER_RUN_OPTS environment variable. Parses the options string into command line arguments.

userid()

@spec userid() :: String.t()

Returns the current user ID by executing id -u.