Pix.Pipeline.SDK (pix v0.0.0)

Provides APIs for building Dockerfiles pipelines programmatically.

The names of the functions map to the Dockerfile commands. Some of the functions have additional options to extend the functionality of the commands (eg. private and cache in stage/2) and some functions are extension to the commands (eg. output/2).

For example:

import Pix.Pipeline.SDK

pipeline =
  pipeline("gohello", description: "My pipeline description")
  |> stage("build", from: "golang:1.23", private: true)
  |> copy("hello.go", ".")
  |> run("go build -o hello hello.go")
  |> stage("app", from: "scratch")
  |> copy("hello", ".", from: "build")
  |> cmd(["/hello"])

pipeline
|> dump()
|> IO.puts()

Output

FROM golang:1.23 AS build
COPY hello.go .
RUN go build -o hello hello.go

FROM scratch AS app
COPY --from="build" hello .
CMD ["/hello"]

Types

args()

@type args() :: %{required(name :: String.t()) => default_value :: nil | String.t()}

command()

@type command() :: String.t()

iargs()

@type iargs() :: [String.t(), ...]

instruction()

@type instruction() :: {command(), options(), iargs()}

options()

@type options() :: Keyword.t()

t()

@type t() :: %Pix.Pipeline.SDK{
  args: [instruction()],
  args_: args(),
  description: String.t(),
  directives: [String.t()],
  dockerignore: [String.t()],
  name: String.t(),
  stages: [Pix.Pipeline.SDK.Stage.t()]
}

Functions

add(pipeline, source, destination, options \\ [])

@spec add(
  t(),
  source :: String.t() | [String.t(), ...],
  destination :: String.t(),
  options()
) :: t()

Adds a ADD instruction.

arg(pipeline, name, default_value \\ nil)

@spec arg(t(), name :: String.t() | atom(), default_value :: nil | String.t()) :: t()

Adds a ARG instruction in a stage scope.

cmd(pipeline, command)

@spec cmd(t(), command :: String.t() | [String.t()]) :: t()

Adds a CMD instruction.

copy(pipeline, source, destination, options \\ [])

@spec copy(
  t(),
  source :: String.t() | [String.t(), ...],
  destination :: String.t(),
  options()
) :: t()

Adds a COPY instruction.

dump(sdk)

@spec dump(t()) :: String.t()

Converts the pipeline into a multistage Dockerfile string representation.

entrypoint(pipeline, command)

@spec entrypoint(t(), command :: String.t() | [String.t()]) :: t()

Adds a ENTRYPOINT instruction.

env(pipeline, envs)

@spec env(t(), envs :: Enumerable.t({String.t() | atom(), String.t()})) :: t()

Adds a ENV instruction.

expose(pipeline, port)

@spec expose(t(), port :: String.t()) :: t()

Adds a EXPOSE instruction.

global_arg(pipeline, name, default_value \\ nil)

@spec global_arg(t(), name :: String.t() | atom(), default_value :: nil | String.t()) ::
  t()

Adds a ARG instruction in the global scope.

healthcheck(pipeline, command, options \\ [])

@spec healthcheck(
  t(),
  command :: nil | String.t() | [String.t(), ...],
  options :: options()
) :: t()

Adds a HEALTHCHECK instruction.

label(pipeline, labels)

@spec label(t(), labels :: Enumerable.t({String.t(), String.t()})) :: t()

Adds a LABEL instruction.

output(pipeline, path)

@spec output(t(), Path.t() | [Path.t(), ...]) :: t()

Declare a stage output artifact.

This function doesn't add any instruction to the Dockerfile. It's used to declare the output of the stage.

pipeline(name, options \\ [])

@spec pipeline(name :: String.t(),
  description: String.t(),
  dockerignore: [String.t()],
  directives: [String.t()]
) :: t()

Creates a new pipeline.

Options

description - pipeline description

dockerignore - list of files and directories to ignore in the context

directives - list of parser directives, ie syntax=docker/dockerfile:1.4

pipeline_ctx()

@spec pipeline_ctx() :: String.t()

Build context name of the pipeline ctx directory.

Pix.Pipeline.SDK.copy("foo.sh", ".", from: Pix.Pipeline.SDK.pipeline_ctx())

run(pipeline, command, options \\ [])

@spec run(t(), command :: String.t() | [String.t(), ...], options()) :: t()

Adds a RUN instruction.

For Here-documents string see sigil_h/2.

shell(pipeline, command)

@spec shell(t(), command :: [String.t(), ...]) :: t()

Adds a SHELL instruction.

sigil_h(str, opts)

@spec sigil_h(str :: String.t(), opts :: Keyword.t()) :: String.t()

A basic Here-documents string sigil.

It expands to:

<<EOT
... your string here ...
EOT

Example:

run(pipeline, ~h"""
  if [ "$X" == "x" ]; then
    echo "x!"
  fi
""")

For more heredoc advance feature simply encode the heredoc yourself.

stage(pipeline, stage_name, options \\ [])

@spec stage(t(), stage_name :: String.t(),
  from: String.t(),
  description: String.t(),
  private: boolean(),
  cache: boolean()
) :: t()

Starts a new stage FROM the given base image.

stage("build", from: "golang:1.23")

the optional private and cache options can be used to control the behavior of the stage:

private: true - the stage will not be accessible as a build target, only from other stages.

cache: false - the stage will not be cached, the stage will be built from scratch every time.

stopsignal(pipeline, signal)

@spec stopsignal(t(), name :: String.t()) :: t()

Adds a STOPSIGNAL instruction.

user(pipeline, user)

@spec user(t(), user :: String.t()) :: t()

Adds a USER instruction.

volume(pipeline, volume)

@spec volume(t(), volume :: String.t() | [String.t(), ...]) :: t()

Adds a VOLUME instruction.

workdir(pipeline, workdir)

@spec workdir(t(), workdir :: String.t()) :: t()

Adds a WORKDIR instruction.