Pix.UserSettings (pix v0.0.0)

Pix user settings.

You can set some global settings in your ~/.config/pix/settings.exs file.

The file should evaluate to a map conform to t/0.

For example:

%{
  env: %{
    "PIX_DEBUG" => "true"
  },
  command: %{
    run: %{
      cli_opts: [
        ssh: true
      ]
    },
    shell: %{
      cli_opts: [
        ssh: true
      ]
    }
  }
}

Types

t()

@type t() :: %Pix.UserSettings{
  command: %{
    run: %{cli_opts: OptionParser.parsed()},
    shell: %{cli_opts: OptionParser.parsed()},
    graph: %{cli_opts: OptionParser.parsed()},
    ls: %{cli_opts: OptionParser.parsed()}
  },
  env: %{required(String.t()) => String.t()}
}

Functions

get()

@spec get() :: t()